[OBEY-CAMPAIGN-8deed8b4] Split service_test.go into 4 files by concern
